+++ src/containers/favorites-tab-container.js
